requirements.txt
beautifulsoup4
PyVirtualDisplay
selenium
face_recognition
colorlog
yapsy
fpdf
gitpython